+++ Sources/KeyboardKit/Layout/KeyboardLayoutProvider.swift
@@ -16,8 +16,8 @@ import Foundation
 /// when the keyboard is launched. You can change or replace
 /// it at any time to customize the keyboard layout.
 ///
-/// KeyboardKit Pro can be used to unlock localized provider
-/// implementations for each ``KeyboardLocale``.
+/// KeyboardKit Pro can be used to unlock localized services
+/// for each ``KeyboardLocale``.
 ///
 /// See <doc:Layout-Article> for more information.
 public protocol KeyboardLayoutProvider: AnyObject {