+++ Sources/KeyboardKit/_Deprecated/Autocomplete+Deprecated.swift
@@ -1,11 +1,3 @@
-//
-//  Autocomplete+Deprecated.swift
-//  KeyboardKit
-//
-//  Created by Daniel Saidi on 2023-12-06.
-//  Copyright © 2023-2024 Daniel Saidi. All rights reserved.
-//
-
 import Foundation
 
 @available(*, deprecated, renamed: "AutocompleteService")