+++ Sources/KeyboardKit/_Deprecated/Callouts+Curve.swift
@@ -1,11 +1,3 @@
-//
-//  Callouts+Curve.swift
-//  KeyboardKit
-//
-//  Created by Daniel Saidi on 2021-01-08.
-//  Copyright © 2021-2024 Daniel Saidi. All rights reserved.
-//
-
 import CoreGraphics
 import SwiftUI